Add Doxygen support and code formatting (#3)
* Add Doxygen support and code formatting

With this commit, I've added support for compiling code documentation for the Device Client via Doxygen. I have also tried my best to go through all of the code and document it as best as possible, while running each file through our linter to format according to the specified standards. To generate the documentation, simply enter the 'doc' folder and run 'doxygen'. Right now, I have added the generated files to the .gitignore to reduce clutter and overall size of the package.