Add logger configuration support and improved setup (#12)
* Add logger configuration support and improved setup

With this commit, I have added support for configuring the logger either through the command line, json config, or both. This allows the user to specify the log level, the log type (JSON or STDOUT), and the log file if they are choosing to log to a file. 

I have also done a bit of work to enhance the README, including changing the build instructions to better support older versions of CMAKE and updating the setup.sh script to provide a better experience.

* Update README and FileUtils

I have updated the readme to utilize a more exact specifier for building the test executable for running unit tests. I have also updated a method signature in FileUtils to use a constant string reference based on a PR comment.