Attempt to create FP directories before provisioning
This commit revises the Fleet Provisioning feature so that we attempt to create required directories and set permissions prior to actually running the provisioning process. This way, we do not mark the device as provisioned when we couldn't actually write the files to disk in a secure way. 

I have also modularized FileUtils a bit to allow us to convert a file permission mask to human readable format, automatically create directories and set permissions, and I have written some unit tests to go along with that.