Add permissions validation and defaults configuration (#28)
* Add permissions validation and defaults configuration

As part of this commit, I have made the following changes:

*Added a method to FileUtils.cpp that determines the current permission level of a file or directory. Also added unit tests for this class.

*Added logging to Main.cpp to let the user know whether a feature is disabled or enabled at runtime

* Setup.sh has been modified to configure the correct folders and permissions by default. When folders and files are copies and/or created, we now set the proper permissions that are expected by the device client. 

* Setup.sh now uses the installing user's home directory to create configuration files, as expected by the device client. For standard users, this will typically be /home/<user>/.aws-iot-device-client. For root, this will be /root/.aws-iot-device-client. 

* Path expansion - when we load in configuration values from JSON or CLI, we now expand any paths that include ~ to get the full path to the user's home directory

* Add Job Handler Configuration - we weren't actually passing the job handler directory through to the jobs feature, so I fixed that during this pass

* Jobs permissions - If we own the specified handler, we only execute it if the permissions are 700

* Add incoming file size check

This commit adds a file size check against the incoming configuration files to prevent against loading an extremely large file into memory. The current limit is set to 5KB.

* Fix formatting issues

Fixing issues raised by our linter for formatting

* Consolidate default permissions values

This commit consolidates all of our specified permissions values as part of the Config so that we can assess them all more easily. This also cleans up the code a bit.

* Attempt to create FP directories before provisioning

This commit revises the Fleet Provisioning feature so that we attempt to create required directories and set permissions prior to actually running the provisioning process. This way, we do not mark the device as provisioned when we couldn't actually write the files to disk in a secure way. 

I have also modularized FileUtils a bit to allow us to convert a file permission mask to human readable format, automatically create directories and set permissions, and I have written some unit tests to go along with that.

Co-authored-by: Matt <mwp@amazon.com>