Add Jobs documentation to README
With this change, I've added a large amount of documentation to the README for how to use the Jobs feature within the Device Client. While working through the documentation, I also found that we weren't actually reading a property from the incoming job document so that change has been implemented as well.