Address empty json config values for path expansion
After running the setup script, some path values may be empty if the user decided to skip over enabling a feature such as fleet provisioning for example. This commit verifies whether the values are empty or not prior to performing a path expansion or including the information in the config which could cause the program to abort.