Add log queue transfer to logging API
By default (prior to this commit), the IoT Device Client would use the StdOut logger by default, and after reading in a configuration specifying a File logger, the client would dump any existing logs to StdOut and then log any new messages to the file. This was undesirable since any logs created prior to reading the configuration end up in a different location than logs created after. 

This commit adds support for transferring a LogQueue from one logging implementation to another. With this change, any logs generated prior to fully reading the configuration will be queued, and will be either written by the existing logger or transferred to a new logger to be written later.