Refactor log messaging for config file parsing
We use the same configuration parser for both the standard configuration file as well as the runtime config that's used for fleet provisioning. While this works fine, we have had a number of reports of concern from beta testers that the log messaging is confusing, leading them to believe that they are missing a required configuration file when the logging says that the runtime config is not found

This commit addresses the issue by introducing variable logging based on whether the config file we are looking for is a runtime configuration or not.