Set OpenSSL version explicitly (#48)
* Refactor log messaging for config file parsing

We use the same configuration parser for both the standard configuration file as well as the runtime config that's used for fleet provisioning. While this works fine, we have had a number of reports of concern from beta testers that the log messaging is confusing, leading them to believe that they are missing a required configuration file when the logging says that the runtime config is not found

This commit addresses the issue by introducing variable logging based on whether the config file we are looking for is a runtime configuration or not.

* Set OpenSSL version explicitly

We were pulling the OpenSSL version via the package management repositories, but we want to explicitly point to version 1.1.1 of OpenSSL for consistency.

* Add LINK_DL flag to link against dl from glibc

For our GitHub actions Ubuntu build, we had to modify our OpenSSL setup to build version 1.1.1. This resulted in an issue when compiling against static versions of the SDK libraries that would cause build failure due to inability to resolve symbols related to dynamically loading libraries at runtime. To solve this problem, I have added a compile flag for linking against dl lib from glibc.

Co-authored-by: Matt <mwp@amazon.com>