Add log limits for job child processes
This commit introduces a change to the job engine that enables a limit against the number of lines of STDOUT and STDERR log lines that we will write to the logs. This prevents DOS attacks (accidental or intentional) that may occur if a job causes a large amount of output to be generated.