Add string sanitization utility
As part of our application security review, a vulnerability was identified in which a job handler could generate logs that are dangerous for our application, including format specifiers that can cause a segmentation fault in the logging API. This PR provides an additional safety check that sanitizes incoming strings from both the JobEngine and our configuration reader to avoid issues related to string formatting. There may be additional areas where sanitization is required.