Add support for local certs and improved abort messaging
This commit addresses an issue where we could not properly detect certificates that were placed in the same directory as the device client, since we were not properly resolving the parent directory of those files. I have also added some additional messaging in the event that we abort the program execution based on feedback from beta evaluation.