Add MQTT retry in initial connection request and handle duplicate job notifications (#67)
* Implement connection attempt retry

In order to improve on the resiliency of the device client, we need to add retry logic for both establishing and maintaining our MQTT connection. This commit implements establishment retry via our exponential retry utility. We still need to implement logic for maintaining the connection if it is dropped.

* Add duplicate notification detection in Jobs feature

This commit adds support for detecting and eliminating duplicate job notifications so that we don't run a job more than once. This is helpful for network loss scenarios, where we will receive duplicate messages after the connection is re-established.

Co-authored-by: Matt <mwp@amazon.com>