Add SDK log config (#92)
* Add SDK log config

This commit adds support for configuring the SDK logging. It provides configuration options on the CLI and via JSON, provides help information output from the executable, adds tests for this functionality, and adds an entry to the CHANGELOG describing the change. I have also modified existing log config variables to deconflict the naming and reduce confusion.

* Create SDK log file default constant

This commit creates a constant that specifies the default log location for the SDK, making it a bit easier to make updates to the default file in the future.

Co-authored-by: Matt <mwp@amazon.com>