Address cross-platform compatibility for ANSI color codes (#94)
Prior to this commit, our setup script did not properly encode colors for our prompts on various platforms. This is causes by variances in the default implementation of 'echo' in different shell types (such as Dash, Bash, etc.). To resolve this problem, I have converted the setup script from using echo to printf which is widely considered to be a better option for compatibility. I have tested on Ubuntu and RHEL instances and it appears that the script is now working as expected.

Co-authored-by: Matt <mwp@amazon.com>