Modify GitHub Actions artifact archival to reduce folder hierarchy (#119)
This commit modifies the existing ci.yml file for our Github actions configuration to reduce the number of parent folders included in the generated archive. Prior to this commit, our files were being archived within build/ and setup/ folders. After this change, the binary and the appspec.yml file for CodeDeploy are placed directly in the root directory of the archive, with CodeDeploy specific files going in their own folder and shared libraries being given their own folder.

Co-authored-by: Matt <mwp@amazon.com>