Improve Jobs feature shutdown stability
This commit addresses in an issue in which the jobs feature would not update a job execution successfully upon job completion if a shutdown request had been delivered to the jobs feature by main.cpp. The issue was resolved by adding support for passing an onSuccess callback function to our exponential backoff retry so that we only shut down the feature once we've made some attempts to update the job execution and have tried to wait for a response from the Jobs service. This commit also improves stability by replacing a group of booleans protected by a lock with atomics.