Improve Jobs feature shutdown stability (#120)
* Improve Jobs feature shutdown stability

This commit addresses in an issue in which the jobs feature would not update a job execution successfully upon job completion if a shutdown request had been delivered to the jobs feature by main.cpp. The issue was resolved by adding support for passing an onSuccess callback function to our exponential backoff retry so that we only shut down the feature once we've made some attempts to update the job execution and have tried to wait for a response from the Jobs service. This commit also improves stability by replacing a group of booleans protected by a lock with atomics.

* Resolve gpp compatibility issues

This commit attempts to resolve some gpp compatibility issues by initializing a variable that was not being initialized.

* Change exponentialBackoff onSuccess callback to onComplete

This commit replaces the onSuccess callback of the exponentialBackoff function with an onComplete callback instead. This way, our check to see if the feature should be shutdown is always executed after our retry attempts as opposed to only when the retry attempt is successful.

* Revise retry signature to provide default onComplete pointer

This commit introduces a default argument for the onComplete callback function for the exponential backoff retry. This allows us to remove the requirement to pass 'nullptr' when invoking the exponential backoff function.

* Revise error messages to aid debugging

This commit revises a number of error messages in the JobsFeature to provide better indication of which subscription may be failing in any particular instance. I have also revised the ClientBaseErrorNotification from SUBSCRIPTION_REJECTED to SUBSCRIPTION_FAILED to provide a clearer explanation of what's actually happening.

* Add logging of specific io error

This commit adds logging of the specific IO error that is passed into the subscribe to next job changed ack handler. This way, the user can more easily debug the issue.

Co-authored-by: Matt <mwp@amazon.com>