Update job to "In Progress" after receive
This commit sets the job execution status to in progress once the job is received and processable by the Device Client. In order to accomplish this, the Job engine code was moved to a separate thread, since its activity is typically long running and will block the main thread from receiving the update job execution response.